Move require_version to a separate utils module
